Add symbolication using addr2line when possible.
This commit adds a DwarfSymbolizer struct that can fill symbols using the
addr2line crate when unstripped binaries are available locally. It adds
a SymbolProvider trait (I should really rename breakpad_symbols::Symbolizer)
and implements it for DwarfSymbolizer and Symbolizer, and then also adds a
MultiSymbolProvider struct that will attempt to symbolize a frame given
a list of SymbolProviders. Finally, minidump_stackwalk is changed to use
a MultiSymbolProvider containing a DwarfSymbolizer and a Symbolizer so that
it can attempt to use local DWARF symbols.