convert to scroll for I/O, rework Minidump to use &[u8] instead of Read.
This patch changes how we read the raw minidump structures. Instead of
declaring them as `#[repr(C, packed)]` and transmuting them from raw bytes,
we now use scroll and `#[derive(Pread)]` which is much more sensible. To
facilitate this we also switch from having Minidump take a `Read` type
to taking a `Deref<[u8]>` type, so we can use an owned `Vec<u8>` or
a borrowed `&[u8]` as well as an `memmap::Mmap` if desired. This additionally
makes it possible to have streams and other derived objects with borrowed
data, although this patch does not make such changes yet.