Use references to data in more places instead of copying. Fixes #23
This change adds lifetimes to the `MinidumpStream` trait and several
of the stream structs so that we can store some of the data they contain as
references to the original minidump data instead of making copies. This
is especially nice for `MinidumpMemory`, which can contain a sizable amount
of memory as bytes.