Avoid undeeded lifetime
Self can directly be of lifetime 'a, constraining it to a different
lifetime does not seem to gain any more flexibility.  This simplifies
the signature and thus use.

I don't think there are backwards compatibility concerns with this,
everything which was possible should still be possible.  The lifetime
of the returned stream is still the same 'a as before.