Fix reading annotation object strings
While most of Crashpad's UTF8 strings are terminated with a NUL byte,
the string variant of annotation objects is not. The Breakpad parser for
CrashpadInfo ignores this, never validates the NUL byte, and doesn't
even implement parsing of annotation objects.

To make intentions more clear, this patch splits UTF8 string parsing
into two methods: one that validates the terminating NUL byte, and one
that does not.