Print CrashpadInfo in minidump_dump
Implements support for MinidumpCrashpadInfo in minidump_dump, mostly
equivalent to Breakpad's reference implementation. Breakpad does not
support annotation objects, which we print in a similar fashion.

Note that the interpretation of the GUIDs is endianness dependent,
which means that byte order of the first three fields is inverted for
little-endian Minidumps independent of the host endianness.