Implement STACK CFI unwinding for x86/x64
This teaches breakpad-symbols to parse and evaluate STACK CFI exprs, and plumbs it through to the amd64/x86 stack walkers in minidump-processor. As a result, cfi-based unwinding now works on linux and x64 windows.

To do this, it introduces a FrameWalker trait in breakpad-symbols that abstracts away the details of mapping string register names to loads/stores from the context.

minidump-processor in turn introduces a CfiStackWalker that implements FrameWalker by generically wrapping up the callee's Context and a fresh caller Context. The Context trait is also beefed up to expose the necessary details.