Sketch out STACK WIN unwinding
This implements about 80% of STACK WIN support in the unwinder.
Unforuntately that extra 20% is nasty heuristics that seem to be
necessary to make it work well, so it's disabled for now.

Still, it's progress.

The main major change here is that STACK WIN unwinding requires access
to the callee's callee's (grand-callee's) parameter_size. So StackFrames
now fetch their parameter_size from the symbolicator, and the unwinder
passes down a grand_callee_frame to be expose that frame's parameter_size
in the FrameWalker.

This also adds more detailed docs for STACK CFI.