Add support for the unloaded modules stream.
This also introduces the notion of an 'extended list', which is a format that several
newer sections implement but we hadn't happened to use yet. Most of this code is just
copy-pasting the normal Module and List code.

This is the minimal impl of #168, but ideally we'd use this info for unwinding/symbolication too