Rearrange the workspace
* Move minidump down into a subdirectory so the workspace is virtual
* Move minidump-processor's binary to its own top-level crate
* Add the Cargo.lock to the repo (remove it from .gitignore)

This more accurately reflects the nature of the repo. The `minidump` crate
is in no way the "dominant" one. As a nice side-effect, this means running
`cargo test` or any other command will now properly apply to all the crates,
instead of just `minidump` or the one you're currently cd'd into.

All crates are also bumped to 0.3.0 to be published, so that we can claim the
minidump-stackwalk name.