give every unwinder the full callee frame
cfi unwining should be using callee.instruction instead of the pc/ip in the context. Rather than
duplicate the logic to specially handle the context frame everywhere, just use the fact that we
keep the right value in the callee and pass it down. To avoid having a million things passed down
I also remove several other arguments that are just in the callee frame already. This adds some
boilerplate for pulling out the callee validity, but I think it's fine.