Massively cleanup the interface for minidump-stackwalk
* Add lots of help messages
* Make trace logging way less noisy
* Introduce defaults for --symbols-tmp and --symbols-cache

symbols-tmp is still ignored, but we now implicitly use std::env::temp_dir via the temp_file library
As the new help message for symbols-tmp explains, a different tmp location is needed to atomically swap
files into the cache so that multiple instances can share the cache.

To improve portability/usability, symbols-cache now defaults to std::env::temp_dir()/rust-minidump-cache.
This is especially valuable on Windows where the equivalent of /tmp is in a hidden folder in the User
directory. I primarily work on Windows so this resolves a large friction point. (rust-minidump is largely
very portable, this was the last papercut I knew of!)

This deterministic default potentially introduces some weird attacks where someone writes their own
fake symbol files into the cache but we don't trust the files anyway so I think it's fine?

See rust's documentation for how to modify std::env::temp_dir (lots of options).