Add experimental support for recovering function args
I ended up hacking this together today to help a coworker debug a weird crash report.
Spent the rest of the day polishing it up to make it usable.

This is hidden behind --recover-function-args flag in minidump stackwalk and only
shows up in human output. (minidump-processor similarly hides it in ProcessorOptions.)