Add support for parsing LinuxMaps stream
This is a dump of /proc/self/maps from the crashing process. Details on the format
included in the comments. Because this info does a similar thing to MinidumpMemoryInfoList
I have also introduced a UnifiedMemoryInfo type which you can use to wrap the two up
and not deal with the differences.

minidump_processor now reads both streams and sets up a UnifiedMemoryInfo, but it doesn't
yet do anything with it. Also did some cleanup on the previous patch that I noticed while
designing the unified API.