Stub out implementations of the stringy linux streams
Fixes #207

Adds implementations for:
* LinuxCpuInfo (/proc/cpuinfo)
* LinuxProcStatus (/proc/self/status)
* LinuxEnviron (/proc/self/environ)

The only new data we collect is LinuxCpuInfo.microcode_version, which gets fed through to
cpu_microcode_version in the json schema. All the code for the other streams is written,
so it should now be trivial to grab whatever key we decide is interesting with only a few
new lines.

LinuxAuxv, LinuxCmdLine and LinuxDsoDebug have been excluded from
this patchset as they all have more custom formats than the KEY SEPARATOR VALUE formats
of the implemented streams.