Make all the linux streams handle non-utf8
This adds a new strings submodule with LinuxOsString and LinuxOsStr types that provide
most of the functionality we need. Ultimately though minidump_processor feeds things that
truly are LinuxOsStrings through String::from_utf8_lossy because we need to do *something*
to output them to json/cli.

This also moves all the linux streams to properly using Cow and lifetimes.