Rework the linux stream API to expose an iterator of key-value pairs
Unlike most streams, these ones are just absolutely full of random values,
so it's not practical for us to pull out the values we consider interesting for
the user of the minidump library.

The current iterator is slightly opinionated in that it strips whitespace and
quotes, so the raw bytes of each stream is also exposed in case that's useful.

Breaking change:
Previous extraction of particular values from the Linux* streams (except for LinuxMaps) has been moved to minidump-processor, where it can fetch exactly the values it wants.