Add a timeout to symbol downloads
This also introduces a --symbol-download-timeout-secs flag to minidump-stackwalk
so it's configurable. The default is 1000 seconds (which for a 300MB .sym file
requires a connection speed of 300KB/s).

Weirdly I think adding the timeout at all seems to unstick the download, but I'm
not sure. Either way you can tune the value to your needs.

Fixes #300