Make symbolication async
These libraries are designed for a workflow where you just spawn multiple
processes, and no forward progress can be made on a minidump until symbolication
comes back with a response. So on the surface async is pointless.

However it's necessary for compiling to wasm, and enables clients to use the
library in a different way, and may enable a rewrite of the symbolication
backend to be based on sentry's infrastructure.