Make --human the default output of minidump-stackwalk
There are 3 different binaries which call themselves minidump-stackwalk

1. breakpad's minidump-stackwalk (human output)
2. mozilla's minidump-stackwalk (json output)
3. rust-minidump's minidump-stackwalk (both)

This implementation (mdsw 3) was primarily designed to replace mdsw 2, so
default to json output was "natural". However mdsw 1 is actually run locally
and in the wild more, so emulating their interface is probably the
better default, even if the json output is the "first-class" output.

This will be part of the 0.9.6 release, and is a fairly major breaking change,
but also fairly trivial to handle downstream -- just pass the --json flag.

Note that minidump-stackwalk 0.9.5 introduced --json even while it was the
default, so you can more easily handle this transition by always specifying
--json or --human instead of relying on the default.

fixes #337