Try to parse the csd_version for OS info if the info can't be found in an obvious place.
In minidumps created by breakpad on Linux, the OS's version number
can only be found in its MinidumpSystemInfo's csd_version field.

This adds in a helper to parse out the OS version number and build
number it turns out there's a chance that we're in that
aforementioned scenario. The original csd_version field is left
intact on the MinidumpSystemInfo in case the value is needed for
any reason by any consumers.