Switch to debugid crate instead of strings (#432)
Switch to debugid crate instead of strings

The Debug ID and Code ID were represented by strings in the breakpad
format.  This switches this to the DebugId and CodeId types from the
debugid crate as also used by symbolic.

The benefit is strong types and that DebugId becomes Copy.


Notes from Gankra:

One side-effect of this is that codeids are now uniformly lowercased, 
where previously they were uppercased. There is no "correct" casing
because they are stored in the minidump as integers. That said, be
careful if you use these strings directly and are using case-sensitive
tooling. (Microsoft's tooling is case-insensitive, in our testing.)