Fix some overflows on x86 and amd64 stack walker (#442)
While stack scanning we would try to recover the frame pointer *even* 
on the first iteration of a scan. This doesn't make sense because if we
think we've recovered the return pointer on the first iteration, we are
concluding that the stack frame is the *size* of just one pointer. Such
a frame obviously cannot contain a second pointer, so we were
actually just reaching into memory outside the stack frame and
potentially the stack itself.

This wasn't dangerous but was obviously incorrect.