Return a nil CodeId when no reasonable value can be derived
Modules in minidumps generated on Linux may omit CV records
entirely if a build ID could not be found for them. When that
happens, the entire module is zeroed-out except for its name,
its image's base memory address and its image's size.

No meaningful code ID could or should be derived from such a
module, and so a nil code ID is returned in this scenario.