Use enum for pointer width instead of integer
Allowing for an arbitrary number of bits in a word means we may have to
write code that is general enough to handle any number of bits, which can
get pretty complicated, and is generally not needed.

Case-in-point: Even though this function has only been used twice, we can
already see it being matched against "Some(4)" so it can be casted to a u32.