fix: Use the correct stack memory for exception context
When a minidump has an exception context that is different from the thread context,
it may refer to a different memory region, which we should use in that case rather
than the stack memory region.

This also removes the confusing/duplicated bounds checking on memory access, as scroll does its own bounds checking when reading the value anyways.