Add support for parsing INLINE_ORIGIN and INLINE.
This parses the new format but produces the same output as before.
So with this patch alone, we will not get any inline information in the
output.
But this patch allows us to switch to the new sym format without
breaking anything (hopefully).

In `fill_symbol` we now call `get_outermost_source_location`, which has
to look at the inline information, because just by looking at the lines
you would not get the same output as before with the new sym files.
That's because, when inline information is present in sym files, the
line records now describe the lines *at the deepest level of inlining*.
In sym files without inline information, the line records describe the
lines at the outermost level.