Detect "null pointer in diguise"
This adds functionality to detect a "null pointer in disguise", which occurs
when a null pointer is accessed with an offset (for example, in the context of
an array or a field within a struct).

When such a situation is encountered, it will report the crash address
as "0" instead of whatever non-zero offset is given by the report.