Lookup debug info using code info
When fetching a symbol file from an HTTP symbol supplier and the debug
filename and debug identifier are missing, try looking up the debug
information using the code info. If the lookup is successful, use that
debug info to fetch the symbol file from the cache and symbol servers.

This uses a new feature in Mozilla Symbols Server where requesting a
download with:

```
/<code_file>/<code_id>/<code_file>.sym
```

will return an HTTP 302 with the download API url using the debug file
and debug id if the Mozilla Symbols Server has a record for that file.

Fixes #870