Fix redirect url parsing (#882)
When the symbol supplier does a codeinfo lookup and it's successful, the
symbol supplier gets back an HTTP 302 with a url to redirect to in the
Location header. The path of that url may have additional parts at the
beginning that should be ignored--the debug file and debug id are the
third and second from the end.

This fixes the path part parsing so
individual_lookup_debug_info_by_code_info picks up the debug file and
debug id correctly.