Interpret Windows error-code correctly even when they're sign-extended.
When interpreting Windows status code that have been stored as 64-bit
values, ignore the upper 32 bits as they're likely to contain bogus
data. In particular windbg.dll seems to sign-extend the 32-bit value
when storing it as a 64-bit one.

Note that we're not really throwing away data with this patch as the
status codes are supposed to be 32 bits, so we shouldn't have relied on
the upper 32 bits being cleared in the first place.

This fixes issue #893