Update framehop and change `object` to have minimal features.
Consumers can add more features to change object reading if desired.

The way the crate is written, at least one of
`coff`/`elf`/`macho`/`pe`/`wasm`/`xcoff` must be enabled for
`object::read::File` to be available. While we could limit the features
even further and make consumers choose which formats to support, this
instead assumes that they want to be able to read any of the formats
since we need to choose at least one anyway.